Remove EIP-3554 from the Espresso Hardfork (#296)
EIP-3198 added the BaseFee opcode to the ethereum virtual machine. Initially that was to be included in the espresso hardfork for the Celo network, but there is not a well formed way to implement the opcode with how the GasPriceMinimum mechanism is implemented.

In particular, the gas price minimum (equivalent to the BaseFee), is updated in the GasPriceMinimum smart contract. The gas price minimum is also stored in the contract and can be accessed. The value returned by OpBaseFee can only be determined by calling into the GasPriceMinimum contract. If the OpBaseFee opcode were to exist, the GasPriceMinimum would not be able to use that opcode as it is self-referential. It is possible to limit the contract to using a subset of the EVM, but that option is tricky and adds a lot of complexity.

Instead of OpBaseFee, smart contract developers can use getGasPriceMinimum in the GasPriceminimum smart contract.